いろいろコミット

+++ Src/Object/UnitBase.cpp
@@ -241,7 +241,7 @@ void UnitBase::SetSpeed(const float percent)
 }
 
 //‘Ì—Í‹­‰»
-void UnitBase::SetHpMax(const float hp)
+void UnitBase::SetHpMax(const int hp)
 {
 	hpMax_ += hp;
 }
